Refactor project structure with include and src folders
