--- Students_entry.cpp
@@ -1,8 +1,0 @@
-#include "Students_entry.h"
-
-std::ostream& operator<<(std::ostream& os, const Students_entry& entry) {
-	os << "ФИО: " << entry.fio << ", "
-		<< "Инструмент: " << entry.instrument << ", "
-		<< "Преподаватель: " << entry.teacher;
-	return os;
-}
