+++ src/structures/Students_entry.cpp
@@ -1,0 +1,8 @@
+#include "structures/Students_entry.h"
+
+std::ostream& operator<<(std::ostream& os, const Students_entry& entry) {
+	os << "ÔÈÎ: " << entry.fio << ", "
+		<< "Èíñòðóìåíò: " << entry.instrument << ", "
+		<< "Ïðåïîäàâàòåëü: " << entry.teacher;
+	return os;
+}
