+++ src/structures/Teacher.cpp
@@ -1,4 +1,4 @@
-#include "Teacher.h"
+#include "structures/Teacher.h"
 
 std::ostream& operator<<(std::ostream& os, const Teacher& teacher) {
     os << teacher.surname << " " << teacher.initials;
